09.other.projects/01.youtube/requirements.txt
moviepy==1.0.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
httplib2>=0.20.0
oauth2client>=4.1.3
pathlib>=1.0.1
typing>=3.7.4 